# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Squawk social media API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current authenticated user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /users/me/profile:
    put:
      operationId: updateMyProfile
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /users/me/onboard:
    post:
      operationId: onboardUser
      tags: [users]
      summary: Complete onboarding (create user record on first sign-in)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "200":
          description: User already existed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /users/{username}:
    get:
      operationId: getUserByUsername
      tags: [users]
      summary: Get user profile by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /users/{username}/posts:
    get:
      operationId: getUserPosts
      tags: [users]
      summary: Get posts by a user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostPage"
  /users/{username}/followers:
    get:
      operationId: getUserFollowers
      tags: [users]
      summary: Get followers of a user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Followers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
  /users/{username}/following:
    get:
      operationId: getUserFollowing
      tags: [users]
      summary: Get users that this user follows
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Following list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
  /users/suggested:
    get:
      operationId: getSuggestedUsers
      tags: [users]
      summary: Get suggested users to follow
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Suggested users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"

  # ─── FOLLOWS ──────────────────────────────────────────────────────────────────
  /follows/{username}:
    post:
      operationId: followUser
      tags: [follows]
      summary: Follow a user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Now following
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: unfollowUser
      tags: [follows]
      summary: Unfollow a user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unfollowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ─── POSTS ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    UserProfile:
      type: object
      required: [id, clerkId, username, displayName, createdAt, followersCount, followingCount, postsCount, isFollowing]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        isVerified:
          type: boolean
        followersCount:
          type: integer
        followingCount:
          type: integer
        postsCount:
          type: integer
        isFollowing:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UserSummary:
      type: object
      required: [id, username, displayName, isFollowing]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        isVerified:
          type: boolean
        isFollowing:
          type: boolean
        bio:
          type: ["string", "null"]
    OnboardInput:
      type: object
      required: [username, displayName]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
        displayName:
          type: string
          minLength: 1
          maxLength: 60
        bio:
          type: string
        avatarUrl:
          type: string
    ProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 60
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
    FollowResult:
      type: object
      required: [isFollowing, followersCount]
      properties:
        isFollowing:
          type: boolean
        followersCount:
          type: integer

    # ─── POST ──────────────────────────────────────────────────────────────
    Post:
      type: object
      required: [id, authorId, author, mediaUrl, mediaType, likesCount, commentsCount, viewsCount, isLiked, isSaved, createdAt]
      properties:
        id:
          type: integer
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/UserSummary"
        caption:
          type: ["string", "null"]
        mediaUrl:
          type: string
        mediaType:
          type: string
          enum: [image, video]
        thumbnailUrl:
          type: ["string", "null"]
        hashtags:
          type: array
          items:
            type: string
        likesCount:
          type: integer
        commentsCount:
          type: integer
        viewsCount:
          type: integer
        isLiked:
          type: boolean
        isSaved:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    PostPage:
      type: object
      required: [posts, hasMore]
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        hasMore:
          type: boolean
        nextCursor:
          type: ["integer", "null"]
...