# .migration-backup/lib/api-zod/src/generated/types/followResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Squawk social media API
 * OpenAPI spec version: 0.1.0
 */

export interface FollowResult {
  isFollowing: boolean;
  followersCount: number;
}